+++ xarray/tests/__init__.py
@@ -21,6 +21,11 @@ except ImportError:
     import unittest
 
 try:
+    from unittest import mock
+except ImportError:
+    import mock
+
+try:
     import scipy
     has_scipy = True
 except ImportError:
@@ -83,6 +88,17 @@ try:
 except ImportError:
     has_rasterio = False
 
+try:
+    import pathlib
+    has_pathlib = True
+except ImportError:
+    try:
+        import pathlib2
+        has_pathlib = True
+    except ImportError:
+        has_pathlib = False
+
+
 # slighly simpler construction that the full functions.
 # Generally `pytest.importorskip('package')` inline is even easier
 requires_matplotlib = pytest.mark.skipif(
@@ -105,6 +121,9 @@ requires_bottleneck = pytest.mark.skipif(
     not has_bottleneck, reason='requires bottleneck')
 requires_rasterio = pytest.mark.skipif(
     not has_rasterio, reason='requires rasterio')
+requires_pathlib = pytest.mark.skipif(
+    not has_pathlib, reason='requires pathlib / pathlib2'
+)
 
 
 try: